Rename files in singular to match class names
